# api/openapi.yaml
openapi: 3.0.3
info:
  title: SimpleTrack Analytics Service API
  version: 0.1.0
  description: HTTP API for event collection, tracker delivery, health checks, and internal readback.
paths:
...
  /tracker.js:
    get:
      summary: Return the browser tracker script.
      responses:
        "200":
          description: Tracker JavaScript asset.
          headers:
            Cache-Control:
              schema:
                type: string
          content:
            application/javascript:
              schema:
                type: string
        "404":
          description: Tracker script is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /collect:
    options:
      summary: CORS preflight for event collection.
      responses:
        "204":
          description: Preflight accepted by Fiber CORS middleware.
    post:
      summary: Accept one analytics event.
      parameters:
        - in: header
          name: X-SimpleTrack-Write-Key
          schema:
            type: string
          required: false
        - in: header
          name: Authorization
          schema:
            type: string
          required: false
          description: Bearer write key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectRequest"
      responses:
        "202":
          description: Event accepted or intentionally filtered before publish.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptedResponse"
        "400":
          description: Invalid JSON or validation failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid write key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Source disabled or origin not allowed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal collect failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/realtime:
    get:
      summary: Read recent accepted events for one source.
      parameters:
        - $ref: "#/components/parameters/WriteKey"
        - $ref: "#/components/parameters/QueryAuthorization"
        - in: query
          name: since
          schema:
            type: string
            format: date-time
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 50
      responses:
        "200":
          description: Recent events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RealtimeResponse"
        "400":
          description: Invalid query parameter or missing write key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, expired, not-yet-valid, or unknown query token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Origin is not allowed, query token scope does not allow Realtime, or source readback policy disables Realtime.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Query API is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/events:
    get:
      summary: List accepted events for one source and time range.
      parameters:
        - $ref: "#/components/parameters/WriteKey"
        - $ref: "#/components/parameters/QueryAuthorization"
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 100
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: event_name
          schema:
            type: string
        - in: query
          name: distinct_id
          schema:
            type: string
        - in: query
          name: visit_id
          schema:
            type: string
        - in: query
          name: sort_field
          schema:
            type: string
            enum: [event_time, received_at, event_name]
        - in: query
          name: sort_direction
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: property_filter
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          description: Repeatable URL-encoded JSON filter payloads.
      responses:
        "200":
          description: Event list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventsResponse"
        "400":
          description: Invalid query parameter, filter, or missing write key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, expired, not-yet-valid, or unknown query token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Origin is not allowed, query token scope does not allow Events, or source readback policy disables Events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Query API is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/goals:
    get:
      summary: Count events matching one P1 Goal rule for one source and time range.
      parameters:
        - $ref: "#/components/parameters/WriteKey"
        - $ref: "#/components/parameters/QueryAuthorization"
        - in: query
          name: event_name
          required: true
          schema:
            type: string
          description: Event name used by the minimal Goal rule.
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Goal event count summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalResponse"
        "400":
          description: Invalid query parameter, missing event name, or missing write key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, expired, not-yet-valid, or unknown query token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Origin is not allowed, query token scope does not allow Goals, or source readback policy disables Goals.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Query API is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/properties:
    get:
      summary: List observed property definitions for one source.
      parameters:
        - $ref: "#/components/parameters/WriteKey"
        - $ref: "#/components/parameters/QueryAuthorization"
        - in: query
          name: scope
          schema:
            type: string
            enum: [event, user]
          description: Optional property scope filter.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
      responses:
        "200":
          description: Source-scoped property catalog.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyCatalogResponse"
        "400":
          description: Invalid query parameter or missing write key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, expired, not-yet-valid, or unknown query token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Origin is not allowed, query token scope does not allow Properties, or source readback policy disables Properties.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Property catalog API is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/kafka/diagnostics:
    get:
      summary: Read process-local Kafka EventBus diagnostics.
      description: Internal operator endpoint. The response is a process-local diagnostic snapshot and does not include broker credentials, raw event payloads, or authoritative broker lag.
      parameters:
        - $ref: "#/components/parameters/QueryAuthorization"
      responses:
        "200":
          description: Kafka EventBus diagnostic snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KafkaDiagnosticsResponse"
        "401":
          description: Missing, expired, not-yet-valid, or unknown query token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Query token scope does not allow Kafka diagnostics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Kafka diagnostics are not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    WriteKey:
      in: query
      name: write_key
      required: true
      schema:
        type: string
    QueryAuthorization:
      in: header
      name: Authorization
      required: true
      schema:
        type: string
      description: Bearer query token.
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    AcceptedResponse:
      type: object
      required: [id, received_at]
      properties:
        id:
          type: string
        received_at:
          type: string
          format: date-time
        filtered:
          type: boolean
    CollectRequest:
      type: object
      required: [event_name, distinct_id]
      properties:
        write_key:
          type: string
        id:
          type: string
        event_name:
          type: string
        distinct_id:
          type: string
        session_id:
          type: string
        visit_id:
          type: string
        event_time:
          type: string
          format: date-time
        properties:
          type: object
          additionalProperties: true
        user_properties:
          type: object
          additionalProperties: true
    SourceResponse:
      type: object
      required: [tenant_id, project_id, source_id, source_type]
      properties:
        tenant_id:
          type: string
        project_id:
          type: string
        source_id:
          type: string
        source_type:
          type: string
    EventItem:
      type: object
      required: [id, tenant_id, project_id, source_id, source_type, event_name, distinct_id, event_time, received_at]
      properties:
        id:
          type: string
        tenant_id:
          type: string
        project_id:
          type: string
        source_id:
          type: string
        source_type:
          type: string
        event_name:
          type: string
        distinct_id:
          type: string
        session_id:
          type: string
        visit_id:
          type: string
        event_time:
          type: string
          format: date-time
        received_at:
          type: string
          format: date-time
        properties:
          type: object
          additionalProperties: true
        user_properties:
          type: object
          additionalProperties: true
        source:
          type: string
    RealtimeResponse:
      type: object
      required: [source, items, since, limit]
      properties:
        source:
          $ref: "#/components/schemas/SourceResponse"
        items:
          type: array
          items:
            $ref: "#/components/schemas/EventItem"
        since:
          type: string
          format: date-time
        limit:
          type: integer
        query_evidence:
          $ref: "#/components/schemas/QueryEvidence"
    EventsResponse:
      type: object
      required: [source, items, limit, offset, from, to]
      properties:
        source:
          $ref: "#/components/schemas/SourceResponse"
        items:
          type: array
          items:
            $ref: "#/components/schemas/EventItem"
        limit:
          type: integer
        offset:
          type: integer
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        query_evidence:
          $ref: "#/components/schemas/QueryEvidence"
    GoalResponse:
      type: object
      required: [source, event_name, from, to, matching_events, has_data]
      properties:
        source:
          $ref: "#/components/schemas/SourceResponse"
        event_name:
          type: string
          description: Event name used by the minimal Goal rule.
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        matching_events:
          type: integer
          format: int64
          description: Exact event count returned by analytics-core.
        has_data:
          type: boolean
          description: True when at least one matching event exists.
        query_evidence:
          $ref: "#/components/schemas/QueryEvidence"
    PropertyCatalogItem:
      type: object
      required: [scope, name, value_type, first_seen_at, last_seen_at]
      properties:
        scope:
          type: string
          enum: [event, user]
        name:
          type: string
        value_type:
          type: string
          enum: [null, string, number, bool]
        first_seen_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
    PropertyCatalogResponse:
      type: object
      required: [source, items, limit]
      properties:
        source:
          $ref: "#/components/schemas/SourceResponse"
        items:
          type: array
          items:
            $ref: "#/components/schemas/PropertyCatalogItem"
        limit:
          type: integer
...
    KafkaDiagnosticsResponse:
      type: object
      required: [topic, dead_letter_topic, worker_pool, completion_gate, commits, metrics]
      properties:
        topic:
          type: string
        dead_letter_topic:
          type: string
        worker_pool:
          $ref: "#/components/schemas/KafkaWorkerPoolDiagnostics"
        completion_gate:
          $ref: "#/components/schemas/KafkaCompletionGateDiagnostics"
        commits:
          type: array
          items:
            $ref: "#/components/schemas/KafkaOrderedCommitDiagnostics"
        paused:
          type: object
          description: Topic-to-partition list currently paused by local backpressure.
          additionalProperties:
            type: array
            items:
              type: integer
              format: int32
        metrics:
          $ref: "#/components/schemas/KafkaMetricsDiagnostics"
...